Move a helper function for getting events to substrateUtils
