+++ utils/e2e-tests/ts/tests/swap/evmToNative.ts
@@ -6,8 +6,7 @@ import { beforeEachWithCleanup } from "../../lib/lifecycle";
 import evmSwap from "../../lib/abis/evmSwap";
 import { decodeEventLog } from "viem";
 import * as substrate from "../../lib/substrate";
-import { getNativeBalance } from "../../lib/substrateUtils";
-import { AnyJson } from "@polkadot/types-codec/types";
+import { getEvents, getNativeBalance } from "../../lib/substrateUtils";
 
 const evmToNativeSwapPrecompileAddress =
   "0x0000000000000000000000000000000000000900";
@@ -16,14 +15,6 @@ const bridgePotNativeAccount =
   "hmpwhPbL5XJM1pYFVL6wRPkUP5gHQyvC6R5jMkziwnGTQ6hFr";
 const feesPotNativeAccount =
   "hmpwhPbL5XJTYPWXPMkacfqGhJ3eoQRPLKphajpvcot5Q5zkk";
-
-type Event = {
-  event: {
-    method: string;
-    section: string;
-    data: AnyJson;
-  };
-};
 
 describe("evm to native tokens swap", () => {
   let node: RunNodeState;
@@ -146,21 +137,13 @@ describe("evm to native tokens swap", () => {
     });
     expect(evmSwapPrecompileBalance).toEqual(0n);
 
-    const blockNumber = swapTxReceipt.blockNumber;
-    const blockHash = await substrateApi.rpc.chain.getBlockHash(blockNumber);
-    const substrateApiAt = await substrateApi.at(blockHash);
-    const events = (await substrateApiAt.query["system"]?.[
-      "events"
-    ]?.()) as unknown as Event[];
+    const events = await getEvents(substrateApi, swapTxReceipt.blockNumber);
 
     events.forEach((item) => {
       const section = item.event.section;
       const method = item.event.method;
       const data = JSON.stringify(item.event.data);
 
-      console.log(section);
-      console.log(method);
-      console.log(data);
       expect([section, method, data]).not.toEqual([
         "evmSystem",
         "NewAccount",
